// services/cCREs-service/graphqlservice/src/postgres/types.ts
export type CCREMaxZEntry = {
    accession: string;
    score: number;
};

export type CellTypeAssayEntry = {
    biosample_name: string;
    [key: string]: string;
};

export type ZScoreEntry = {
    experiment_accession: string;
    rdhs: string;
    score: number;
};

export type ZScoreParameters = {
    assembly: string;
    experiment?: string[];
    rDHS?: string[];
    minimum_score?: number;
    maximum_score?: number;
};

export type ZScoreHistogramParameters = {
    histogram_minimum: number;
    histogram_maximum: number;
    histogram_bins: number;
};